fix(msp): use the Web Worker to process topology metadata
